feat: Added user identifier after login
