+++ app/(ctabs)/_layout.tsx
@@ -51,7 +51,9 @@ export default function TabLayout() {
         <Tabs.Screen 
             name="reminders"  
             options={{
-                tabBarIcon: ({size, color }) => <MaterialCommunityIcons name="alarm" size={size} color={color} />, 
+tabBarIcon: ({ size, color }) => (
+  <MaterialCommunityIcons name="heart-outline" size={size} color={color} />
+)
             }}
         />
         <Tabs.Screen
