+++ app/index.tsx
@@ -3,5 +3,5 @@ import React from 'react'
 import { Redirect } from 'expo-router'
 
 export default function Index() {
-  return <Redirect href="/(tabs)" />
+  return <Redirect href="/(auth)/login" />
 }
